+++ src/Practitioner/MobileNav.js
@@ -2,12 +2,10 @@ import React,{useEffect} from 'react'
 import { makeStyles } from '@material-ui/core/styles';
 import BottomNavigation from '@material-ui/core/BottomNavigation';
 import BottomNavigationAction from '@material-ui/core/BottomNavigationAction';
-import AddCircleOutlineIcon from '@material-ui/icons/AddCircleOutline';
 import HomeIcon from '@material-ui/icons/Home'
 import ForumIcon from '@material-ui/icons/Forum';
 import InfoIcon from '@material-ui/icons/Info';
 import { inject, observer } from 'mobx-react';
-import EventAvailableIcon from '@material-ui/icons/EventAvailable';
 import useStores from '../Basics/UseStores';
 import Colors from '../Basics/Colors';
 import Badge from '@material-ui/core/Badge'
